+++ addons/goggles/CfgGesturesMale.hpp
@@ -1,0 +1,9 @@
+class CfgGesturesMale {
+    class States {
+        class GestureFreezeStand;
+        class GestureWipeFace: GestureFreezeStand {
+            file = QPATHTOF(anim\WipeGlasses.rtm);
+            canPullTrigger = 0;
+        };
+    };
+};
