+++ addons/goggles/CfgMovesBasic.hpp
@@ -1,0 +1,5 @@
+class CfgMovesBasic {
+    class ManActions {
+        GestureWipeFace[] = {"GestureWipeFace", "gesture"};
+    };
+};
