+++ addons/goggles/CfgWeapons.hpp
@@ -1,0 +1,13 @@
+class CfgWeapons {
+    class H_HelmetB;
+
+    class H_CrewHelmetHeli_B: H_HelmetB {
+        ACE_Protection = 1;
+    };
+    class H_PilotHelmetHeli_B: H_HelmetB {
+        ACE_Protection = 1;
+    };
+    class H_PilotHelmetFighter_B: H_HelmetB {
+        ACE_Protection = 1;
+    };
+};
