+++ addons/goggles/ComplexEffects.hpp
@@ -1,0 +1,9 @@
+class RifleAssaultCloud {
+    ACE_Goggles_BulletCount = 4;
+};
+class MachineGunCloud {
+    ACE_Goggles_BulletCount = 3;
+};
+class SniperCloud {
+    ACE_Goggles_BulletCount = 1;
+};
